+++ ikarus/finiteelements/mechanics/assumedstress.hh
@@ -193,6 +193,11 @@ protected:
     using ScalarType = Traits::ctype;
     asApplicabilityCheck();
     auto correctbeta = [&]<typename AssumedStressT>(const AssumedStressT& asFunction) {
+      if (correction.size() != par.globalSolution().size())
+        DUNE_THROW(Dune::NotImplemented,
+                   "Solution vector and correction vector should be of the same size. Check if DBCOption::Full is "
+                   "used. The sizes are " +
+                       std::to_string(par.globalSolution().size()) + " and " + std::to_string(correction.size()));
       const auto localdxBlock = Ikarus::FEHelper::localSolutionBlockVector<Traits, Eigen::VectorXd, double>(
           correction, underlying().localView());
       const auto localdx = Dune::viewAsFlatEigenVector(localdxBlock);